+++ tests/unit/test_policy_shards_effect.py
@@ -136,7 +136,7 @@ POLICY_SHARDS_EFFECT_SCENARIOS = {
             ),
             PolicyShard(
                 effect="Deny",
-                effective_action=EffectiveAction(inclusion=Action("s3:PutObject"), exclusions=frozenset()),
+                effective_action=EffectiveAction(inclusion=Action("s3:*"), exclusions=frozenset()),
                 effective_resource=EffectiveResource(
                     inclusion=Resource("*"), exclusions=frozenset({Resource("arn:aws:s3:::examplebucket/*")})
                 ),
@@ -152,9 +152,7 @@ POLICY_SHARDS_EFFECT_SCENARIOS = {
         "expected": [
             PolicyShard(
                 effect="Allow",
-                effective_action=EffectiveAction(
-                    inclusion=Action("s3:*"), exclusions=frozenset({Action("s3:PutObject")})
-                ),
+                effective_action=EffectiveAction(inclusion=Action("s3:*"), exclusions=frozenset()),
                 effective_resource=EffectiveResource(
                     inclusion=Resource("arn:aws:s3:::examplebucket/*"), exclusions=frozenset()
                 ),
@@ -192,50 +190,6 @@ POLICY_SHARDS_EFFECT_SCENARIOS = {
                 conditions=frozenset(),
                 not_conditions=frozenset(),
             ),
-            # PolicyShard(
-            #     effect="Allow",
-            #     effective_action=EffectiveAction(
-            #         inclusion=Action("s3:*"), exclusions=frozenset({Action("s3:PutObject")})
-            #     ),
-            #     effective_resource=EffectiveResource(
-            #         inclusion=Resource("arn:aws:s3:::examplebucket/*"), exclusions=frozenset()
-            #     ),
-            #     effective_principal=EffectivePrincipal(
-            #         inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
-            #     ),
-            #     conditions=frozenset(
-            #         {Condition(key="aws:PrincipalOrgId", operator="StringNotEquals", values=["o-123456"])}
-            #     ),
-            #     not_conditions=frozenset(),
-            # ),
-            # PolicyShard(
-            #     effect="Allow",
-            #     effective_action=EffectiveAction(inclusion=Action("s3:PutObject"), exclusions=frozenset()),
-            #     effective_resource=EffectiveResource(
-            #         inclusion=Resource("*"), exclusions=frozenset({Resource("arn:aws:s3:::examplebucket/*")})
-            #     ),
-            #     effective_principal=EffectivePrincipal(
-            #         inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
-            #     ),
-            #     conditions=frozenset(
-            #         {Condition(key="aws:PrincipalOrgId", operator="StringNotEquals", values=["o-123456"])}
-            #     ),
-            #     not_conditions=frozenset(
-            #         {Condition(key="s3:x-amz-server-side-encryption", operator="StringNotEquals", values=["AES256"])}
-            #     ),
-            # ),
-            # PolicyShard(
-            #     effect="Allow",
-            #     effective_action=EffectiveAction(inclusion=Action("s3:GetObject"), exclusions=frozenset()),
-            #     effective_resource=EffectiveResource(
-            #         inclusion=Resource("arn:aws:s3:::examplebucket/*"), exclusions=frozenset()
-            #     ),
-            #     effective_principal=EffectivePrincipal(
-            #         inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
-            #     ),
-            #     conditions=frozenset(),
-            #     not_conditions=frozenset(),
-            # ),
         ],
     },
     "identical_except_different_conditions": {
@@ -280,50 +234,6 @@ POLICY_SHARDS_EFFECT_SCENARIOS = {
                     {Condition(key="s3:x-amz-server-side-encryption", operator="StringNotEquals", values=["AES256"])}
                 ),
             ),
-            # PolicyShard(
-            #     effect="Allow",
-            #     effective_action=EffectiveAction(inclusion=Action("s3:*"), exclusions=frozenset()),
-            #     effective_resource=EffectiveResource(inclusion=Resource("*"), exclusions=frozenset()),
-            #     effective_principal=EffectivePrincipal(
-            #         inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
-            #     ),
-            #     conditions=frozenset(
-            #         {Condition(key="aws:PrincipalOrgId", operator="StringNotEquals", values=["o-123456"])}
-            #     ),
-            #     not_conditions=frozenset(
-            #         {Condition(key="s3:x-amz-server-side-encryption", operator="StringNotEquals", values=["AES256"])}
-            #     ),
-            # ),
-            # PolicyShard(
-            #     effect="Allow",
-            #     effective_action=EffectiveAction(inclusion=Action("s3:*"), exclusions=frozenset()),
-            #     effective_resource=EffectiveResource(
-            #         inclusion=Resource("*"), exclusions=frozenset({Resource("arn:aws:s3:::examplebucket/*")})
-            #     ),
-            #     effective_principal=EffectivePrincipal(
-            #         inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
-            #     ),
-            #     conditions=frozenset(
-            #         {Condition(key="aws:PrincipalOrgId", operator="StringNotEquals", values=["o-123456"])}
-            #     ),
-            #     not_conditions=frozenset(),
-            # ),
-            # PolicyShard(
-            #     effect="Allow",
-            #     effective_action=EffectiveAction(inclusion=Action("s3:*"), exclusions=frozenset()),
-            #     effective_resource=EffectiveResource(
-            #         inclusion=Resource("*"), exclusions=frozenset({Resource("arn:aws:s3:::examplebucket/*")})
-            #     ),
-            #     effective_principal=EffectivePrincipal(
-            #         inclusion=Principal(type="AWS", value="*"), exclusions=frozenset()
-            #     ),
-            #     conditions=frozenset(
-            #         {Condition(key="aws:PrincipalOrgId", operator="StringNotEquals", values=["o-123456"])}
-            #     ),
-            #     not_conditions=frozenset(
-            #         {Condition(key="s3:x-amz-server-side-encryption", operator="StringNotEquals", values=["AES256"])}
-            #     ),
-            # ),
         ],
     },
 }